Change the way the 'odoo' user is created
It is now created in the docker-entrypoint.sh which allow to configure
its ID. The host that runs a container can now configure the id of this
user with the `LOCAL_USER_ID` environment variable, allowing to create
files on a host volume with the permissions of an actual user of the
host's system.

See https://denibertovic.com/posts/handling-permissions-with-docker-volumes/

There is also a new volume `/var/log/odoo` in which we may or not store
the logs. A possible configuration for outputing logs on the host system
is to:

* share a volume "/var/log/odoo:/var/log/odoo"
* set the LOCAL_USER_ID to the desired id of the host system
* set the LOGFILE variable to "/var/log/odoo/odoo.log"