Wait for postgres to really start up
Postgres opens the port but might sometimes still be starting.  It
happens a lot on Travis, when we try to connect on psql, we get "psql:
FATAL:  the database system is starting up".

It will also speed the start of the container, as it will usually have
to wait way less than 1 second.