Remove chown on the content of /opt/odoo
The source code does not need to be writable and the chown is utterly
annoying when we share volumes with the host for development. We could
specify LOCAL_USER_ID on our dev machines, but as this should be done
manually on every machine, I prefer to have a configuration that works
by default.