Add custom scripts in /opt/odoo/start-entrypoint.d
Any script in any language placed in `/opt/odoo/start-entrypoint.d` will be
executed just between the migration and the start of Odoo. The order of
execution of the files is determined by the `run-parts` 's rules.

The database is guaranteed to exist at this point so you can run queries on it.
The scripts are run only if the command is `odoo`/`odoo.py`.

You can add your own scripts in this directory. They must be named something
like `010_abc` (`^[a-zA-Z0-9_-]+$`) and must have no extension (or it would not
be picked up by `run-parts`).