Fix permission issue when installing runtests db
By consistency, in the commit 0c58332 I changed the path to the odoo-bin
script in both places it was used.
However, as long as 'coverage run' accepts a file which doesn't to be
executable, the first call is executed. We could call python2 or python3
on ${ODOO_BIN_PATH}, but the "odoo" script is there for that. So let's
keep "odoo" for the install and use the "odoo-bin" only for 'coverage
run' where it's required.

For the record, the 'odoo-bin' script is eXecutable in the git
repository, but in some cases we optimize the build of the image by
downloading a tar, and it seems the x flag is absent there.