Always set env var `ODOO_REPORT_URL`
By default report.url fallbacks to web.base.url.

As the reports are very workers consuming (needs 3 workers),
we thought that it would be better if they hit the public url
to round robin on the hosts if you have more than one.

This is true but many times does not work because of networks issues,
caching issues, etc (like on minions for instance).

With this change we force it always to be `localhost:8069`.