[FIX] before entrypoint hooks not called
In the new image layout, the `before-migrate-entrypoint.d` and
`start-entrypoint.d` directories are no longer in the root directories
but in the home directory of `odoo`, and the current code in the
entrypoint would not find them, and therefore would not call them.

Fix the path in the docker entrypoint script.