[FIX] disable openblas threads
When numpy is imported, it will load some dependencies, if they are
available. One of these is openblas, which will start some threads,
which we will not be using, but which are consuming precious memory
resources. By setting the OPENBLAS_NUM_THREADS environment variable to
1, we disable the creation of these threads and therefore reduce the
memory usage of the Odoo workers.

This fixes the memory consumption issue when ninja dashboard is
installed.